@@ -14,16 +14,18 @@
   },
   "homepage": "https://github.com/neyko5/balistos-react#readme",
   "dependencies": {
-    "axios": "^0.15.2",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-redux": "^4.4.5",
-    "react-router": "^2.0.0",
-    "react-slider": "^0.7.0", 
+    "axios": "^0.16.1",
+    "prop-types": "^15.5.6",
+    "react": "^15.5.0",
+    "react-dom": "^15.5.0",
+    "react-redux": "^4.4.8",
+    "react-router": "^4.0.0",
+    "react-router-dom": "^4.0.0",
+    "react-slider": "^0.7.0",
     "react-youtube": "^7.3.0",
     "redux": "^3.6.0",
-    "redux-saga": "^0.12.0",
-    "socket.io-client": "^1.7.3",
+    "redux-saga": "^0.14.4",
+    "socket.io-client": "^1.7.2",
     "video-time": "^1.1.0"
   },
   "devDependencies": {
