@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { createStore, applyMiddleware } from 'redux';
-import { Router, Route, browserHistory } from 'react-router';
+import { BrowserRouter as Router, Route } from 'react-router-dom';
 import createSagaMiddleware from 'redux-saga';
 
 import Home from './components/Home/Home';
@@ -24,9 +24,11 @@ sagaMiddleware.run(rootSaga);
 
 ReactDOM.render(
   <Provider store={store}>
-    <Router history={browserHistory}>
-      <Route path="/" component={Home} />
-      <Route path="playlist/:playlist_id" component={Main} />
+    <Router>
+      <div>
+        <Route path="/" component={Home} />
+        <Route path="playlist/:playlist_id" component={Main} />
+      </div>
     </Router>
   </Provider>,
   document.getElementById('root'),
